requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
kfp
pandas
scikit-learn
xgboost
pytest
